Fixed JobGeneratorSystem and added simulationClock, TODO create mouse raycast system

+++ Assets/Scripts/DroNeS/Components/JobCreationTime.cs
@@ -4,6 +4,6 @@ namespace DroNeS.Components
 {
     public struct JobCreationTime : IComponentData
     {
-        public float Value;
+        public double Value;
     }
 }
